+++ project/simulation/src/main/java/datastructures/Result.java
@@ -4,21 +4,21 @@ import java.util.Objects;
 
 public class Result {
     public final int value;
-    public final int requestID;
+    public final long requestID;
     public final CodeExecutionRequest.Operation op;
 
     public int getValue() {
         return value;
     }
 
-    public int getRequestID() {
+    public long getRequestID() {
         return requestID;
     }
 
     public CodeExecutionRequest.Operation getOperation(){ return op;}
 
 
-    public Result(int value, int requestID, CodeExecutionRequest.Operation op) {
+    public Result(int value, long requestID, CodeExecutionRequest.Operation op) {
         this.value = value;
         this.requestID = requestID;
         this.op = op;
